modified initial behavior
