+++ common.blocks/alternative-translation/alternative-translation.bemtree.js
@@ -1,8 +1,8 @@
 block('alternative-translation')(
     content()(function(node) {
-        const translations = node.data.translations || [];
+        const translations = node.data || [];
 
-        return translations.map(translation => {
+        return Array.isArray(translations) && translations.map(translation => {
             return {
                 elem: 'translation',
                 content: translation.target
