+++ server/helpers/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    error: require('./error'),
+    github: require('./github'),
+    translator: require('./translator')
+};
