+++ server/routes.js
@@ -17,16 +17,10 @@ router
         res.redirect(req.session.retpath || '/');
     });
 
-// GitHub routes
 router
-    .get('/', keepRetpath, controllers.github.getContent) // Example: ?doc=https://github.com/bem/bem-method/blob/bem-info-data/articles/bem-for-small-projects/bem-for-small-projects.ru.md
-    .post('/', ensureAuthenticated, controllers.github.createPullRequest);
-
-// Translator routes
-router
-    .get('/tm', controllers.translator.getMemory)
-    .post('/tm', ensureAuthenticated, controllers.translator.saveMemory)
-    .get('/translate', controllers.translator.getTranslate)
-    .get('/alt', controllers.translator.getAltTranslate);
+    .get('/', keepRetpath, controllers.getContent) // Example: ?doc=https://github.com/bem/bem-method/blob/bem-info-data/articles/bem-for-small-projects/bem-for-small-projects.ru.md
+    .post('/send', ensureAuthenticated, controllers.createPullRequest)
+    .post('/translate', controllers.getTranslate)
+    .post('/save', ensureAuthenticated, controllers.saveMemory);
 
 module.exports = router;
